Saving this for future updates
- Added option for different configs between production / development. These are
used in package.json scripts.

This is now working in production but SecurityConfig needs to be updated in future.
After successful login user gets redirected to same view but with logged in status.
It works for me but this routing needs to change so no there is no hardcoded values.